split code up
